708: Create user-facing utils module
